# playwright_2/frontend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Books
  description: Endpoints for managing book availability and lending
paths:
  /books/lend/{id}:
    post:
      tags:
      - Books
      summary: Lend a book
      description: Marks the book with the given ID as lent.
      operationId: lend
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Book lent successfully
        "404":
          description: Book with the given ID not found
        "400":
          description: Book is already lent or unavailable
  /books/search:
    get:
      tags:
      - Books
      summary: Search books by title
      description: Returns books whose titles contain the given query string (case-insensitive).
      operationId: searchBooks
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of matching books returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDto"
  /books/lent:
    get:
      tags:
      - Books
      summary: Get lent books
      description: Retrieves a list of books that are currently lent out.
      operationId: lent
      responses:
        "200":
          description: List of lent books returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDto"
  /books/available:
    get:
      tags:
      - Books
      summary: Get available books
      description: Retrieves a list of all currently available books in the library.
      operationId: available
      responses:
        "200":
          description: List of available books returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookDto"
components:
  schemas:
    BookDto:
      type: object
      description: Represents a book in the library
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the book
          example: 42
        title:
          type: string
          description: Title of the book
          example: The Hitchhiker's Guide to the Galaxy
        lent:
          type: boolean
          description: Indicates whether the book is currently lent out
          example: false